requirements.txt
Django==5.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-grappelli==3.0.5
Pillow==10.2.0
python-dotenv==1.0.1 